Add test, had to move the extractWhiteboard to avoid require.context is not a function error
