--- test/simple_test.js
@@ -1,5 +1,0 @@
-describe('simple', () => {
-  it('should be sane', () => {
-    expect(false).not.toBe(true);
-  });
-});
